add copyright statements to all files; rename suite_test.go; extract ginkgo tests from other unit tests
